+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * Registration module: log to file.
 * Use .ome.tif file extension for output files instead of using .tif extension (but saving in ome-tif format).
 * Place the "filename must include" and "filename must NOT include" filters in a collapsible widget (FileListWidget and FileTableWidget2).
+* New viewer for images, masks and/or cell tracking graphs.
 
 ### Removed